Add files via upload
Hi Reader!

This is my first upload so I am hoping this goes thru correctly.

So, attached are the following:

1. Multi_Year Module 1 - Shows the computation of the main table
2. Multi_Year Module 2 - Shows the computation of the summary table (challenge section)
3. 2014 Screenshot
4. 2015 Screenshot
5. 2016 Screenshot

How did I design the code?

To organize the codes correctly, I divided the scripts into two: Module 1 and Module 2.

Module 1 shows the scripts for the main table. I have it further divided into subsections. I made a Parent Macros file which is triggered by a button click. The Parent Macros file then calls the subsections one by one. I did it this way so it will be more organized and easier to debug as well as less taxing in running my codes.

Module 2 shows the scripts for the challenge part -- the summary table. The codes are rather straightforward for this one.

I hope I was able to explain this clearly. Please reach out to me in case there are issues.